requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-mongodb
langchain-community
pymongo
python-dotenv
google-generativeai
sentence-transformers
torch  # required by sentence-transformers (CPU-only is fine)